Refactor: Rename directories and file names for uniformity
